db-refactor to remove duplication & script updated
