+++ data_generation/data_script.js
@@ -1,5 +1,22 @@
-module.exports = (min, max) => {
+const mongoose = require('mongoose');
+const { Photos } = require('../db/schema.js');
+
+mongoose.connect('mongodb://localhost/photos', { useNewUrlParser: true });
+
+
+const dataScript = (min, max) => {
   const watchPhotos = [];
+
+  watchPhotos.push(
+    {
+      _id: 100,
+      frontImg: 'https://s3-us-west-1.amazonaws.com/hrsf107-the-event-handlers-mvmt-photos/100_back.jpg',
+      sideImg: 'https://s3-us-west-1.amazonaws.com/hrsf107-the-event-handlers-mvmt-photos/100_side.jpg',
+      backImg: 'https://s3-us-west-1.amazonaws.com/hrsf107-the-event-handlers-mvmt-photos/100_back.jpg',
+      box: 'https://s3-us-west-1.amazonaws.com/hrsf107-the-event-handlers-mvmt-photos/box.jpg',
+      styleImg: 'https://s3-us-west-1.amazonaws.com/hrsf107-the-event-handlers-mvmt-photos/100_style.jpg',
+    },
+  );
 
   for (let i = min; i < max; i += 1) {
     const randomPhotoNum = () => {
@@ -10,7 +27,7 @@ module.exports = (min, max) => {
 
     watchPhotos.push(
       {
-        wid: i,
+        _id: i,
         frontImg: `https://s3-us-west-1.amazonaws.com/hrsf107-the-event-handlers-mvmt-photos/${randomPhotoNum()}_back.jpg`,
         sideImg: `https://s3-us-west-1.amazonaws.com/hrsf107-the-event-handlers-mvmt-photos/${randomPhotoNum()}_side.jpg`,
         backImg: `https://s3-us-west-1.amazonaws.com/hrsf107-the-event-handlers-mvmt-photos/${randomPhotoNum()}_back.jpg`,
@@ -21,3 +38,11 @@ module.exports = (min, max) => {
   }
   return watchPhotos;
 };
+
+const data = dataScript(101, 199);
+
+data.forEach((record) => {
+  Photos.create(record, (err) => {
+    if (err) throw err;
+  });
+});
